+++ source/editor/editores.cpp
@@ -7,7 +7,7 @@
 #include <windows.h>
 
 BOOL APIENTRY DllMain( HANDLE hModule, 
-                       DWORD  ul_reason_for_call, 
+                       ULONG  ul_reason_for_call, 
                        LPVOID lpReserved
 					 )
 {
